Improved Steam transport
The existing SteamP2P transport has a number of issues where in it cannot be used for Client Server architectures, includes a version of Steamworks.NET which is not necessarily up to date and makes use of the SteamManager (example not production code).

This change corrects these aspects making this transport suitable for Client Server and Peer 2 Peer, usable by any version of Steamworks.NET from v14 forward, usable by users of custom implementations of Steamworks.NET and 3rd party systems such as Heathen's own Steamworks V2 and this change is not dependent on SteamManager.