Add Option to configure websocket path
This path was hardcoded to /netcode which is not neccessary for
websocket connections, and is an oppinionated choice that breaks in
certain load balancer configurations such as using domains with certs
for WSS connections.

Web sockets are not required to listen on any path at all. This allows
users to delete the path but the default does not break backwards
compatability.